# api/openapi.yaml
openapi: 3.1.0
info:
  title: TallyKeep
  version: 0.1.0-dev
paths:
...
  /api/v1/custodial-providers/{provider_id}:
    get:
      tags:
      - custodial-providers
      summary: Get Provider
      operationId: get_provider_api_v1_custodial_providers__provider_id__get
      parameters:
      - name: provider_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Provider Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustodialProviderOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - custodial-providers
      summary: Patch Provider
      operationId: patch_provider_api_v1_custodial_providers__provider_id__patch
      parameters:
      - name: provider_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Provider Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchCustodialProviderRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustodialProviderOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/custodial-providers/{provider_id}/refresh:
    post:
      tags:
      - custodial-providers
      summary: Refresh Provider
      operationId: refresh_provider_api_v1_custodial_providers__provider_id__refresh_post
      parameters:
      - name: provider_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Provider Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshOut'
        '404':
          description: Provider not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/custodial-providers/{provider_id}/balance:
    get:
      tags:
      - custodial-providers
      summary: Provider Balance
      operationId: provider_balance_api_v1_custodial_providers__provider_id__balance_get
      parameters:
      - name: provider_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Provider Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/custodial-providers/{provider_id}/verify-whitelist:
    get:
      tags:
      - custodial-providers
      summary: Verify Whitelist
      operationId: verify_whitelist_api_v1_custodial_providers__provider_id__verify_whitelist_get
      parameters:
      - name: provider_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Provider Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WhitelistVerificationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/internal/custodial/{provider_id}/poll-cycle:
    post:
      tags:
      - internal
      summary: Poll Cycle
      operationId: poll_cycle_api_v1_internal_custodial__provider_id__poll_cycle_post
      parameters:
      - name: provider_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Provider Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollCycleOut'
        '423':
          description: Secret store is locked
        '404':
          description: Provider not found or archived
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BalanceOut:
      properties:
        provider_id:
          type: string
          format: uuid
          title: Provider Id
        balance_sats:
          type: integer
          title: Balance Sats
        last_polled_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Polled At
      type: object
      required:
      - provider_id
      - balance_sats
      - last_polled_at
      title: BalanceOut
...
    CustodialProviderOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        holding_id:
          type: string
          format: uuid
          title: Holding Id
        provider_kind:
          $ref: '#/components/schemas/ProviderKind'
        display_name:
          type: string
          title: Display Name
        adapter_id:
          type: string
          title: Adapter Id
        can_read:
          type: boolean
          title: Can Read
        can_withdraw:
          type: boolean
          title: Can Withdraw
        whitelist_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Whitelist Address
        whitelist_address_descriptor_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Whitelist Address Descriptor Id
        whitelist_verified:
          type: boolean
          title: Whitelist Verified
        is_active:
          type: boolean
          title: Is Active
        last_polled_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Polled At
        last_error:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Error
        last_known_balance_sats:
          anyOf:
          - type: integer
          - type: 'null'
          title: Last Known Balance Sats
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - holding_id
      - provider_kind
      - display_name
      - adapter_id
      - can_read
      - can_withdraw
      - whitelist_address
      - whitelist_address_descriptor_id
      - whitelist_verified
      - is_active
      - last_polled_at
      - last_error
      - last_known_balance_sats
      - created_at
      - updated_at
      title: CustodialProviderOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PatchCustodialProviderRequest:
      properties:
        display_name:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Display Name
        api_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Api Key
        api_secret:
          anyOf:
          - type: string
          - type: 'null'
          title: Api Secret
        api_passphrase:
          anyOf:
          - type: string
          - type: 'null'
          title: Api Passphrase
      additionalProperties: false
      type: object
      title: PatchCustodialProviderRequest
...
    PollCycleOut:
      properties:
        rows_added:
          type: integer
          title: Rows Added
        rows_updated:
          type: integer
          title: Rows Updated
        rows_unchanged:
          type: integer
          title: Rows Unchanged
        last_polled_at:
          type: string
          format: date-time
          title: Last Polled At
        connection_status:
          type: string
          title: Connection Status
      type: object
      required:
      - rows_added
      - rows_updated
      - rows_unchanged
      - last_polled_at
      - connection_status
      title: PollCycleOut
    ProviderKind:
      type: string
      enum:
      - exchange
      - broker
      - p2p_venue
      title: ProviderKind
...
    RefreshOut:
      properties:
        job_id:
          type: string
          format: uuid
          title: Job Id
      type: object
      required:
      - job_id
      title: RefreshOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WhitelistVerificationOut:
      properties:
        provider_id:
          type: string
          format: uuid
          title: Provider Id
        address:
          type: string
          title: Address
        is_whitelisted:
          type: boolean
          title: Is Whitelisted
        provider_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider Label
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - provider_id
      - address
      - is_whitelisted
      title: WhitelistVerificationOut